refactor(selector4nix): move error utility to top-level module
