+++ src/api/error.rs
@@ -1,7 +1,7 @@
 use axum::response::{IntoResponse, Response};
 use http::StatusCode;
 
-use crate::application::{AppError, AppErrorKind};
+use crate::{AppError, AppErrorKind};
 
 impl IntoResponse for AppError {
     fn into_response(self) -> Response {
