+++ src/application/nar_file/error_impl.rs
@@ -1,5 +1,5 @@
-use crate::application::{AppError, AppErrorKind};
 use crate::domain::nar_file::StreamNarFileError;
+use crate::{AppError, AppErrorKind};
 
 impl From<StreamNarFileError> for AppError {
     fn from(error: StreamNarFileError) -> Self {
