+++ src/application/nar_file/usecase/streaming.rs
@@ -1,10 +1,10 @@
 use std::sync::Arc;
 
 use crate::application::nar_file::actor::{NarFileActorRegistry, NarFileRequest};
-use crate::application::{AppErrorKind, AppOptionExt, AppResult, AppResultExt};
 use crate::domain::common::passthrough_headers::PassthroughHeaders;
 use crate::domain::nar_file::model::NarFileKey;
 use crate::domain::nar_file::port::NarStreamData;
+use crate::{AppErrorKind, AppOptionExt, AppResult, AppResultExt};
 
 pub struct NarFileStreamingUseCase {
     nar_file_registry: Arc<NarFileActorRegistry>,
