+++ src/lib.rs
@@ -5,3 +5,7 @@ pub mod api;
 pub mod application;
 pub mod domain;
 pub mod infrastructure;
+
+mod error;
+
+pub use error::{AppError, AppErrorKind, AppOptionExt, AppResult, AppResultExt};
